Refactor into a Perforator service
