+++ chappy/perforator/src/quic_utils.rs
@@ -1,19 +1,5 @@
-use quinn::{ClientConfig, ServerConfig};
-use std::{error::Error, sync::Arc};
-
-// /// Builds default quinn client config and trusts given certificates.
-// ///
-// /// ## Args
-// ///
-// /// - server_certs: a list of trusted certificates in DER format.
-// fn configure_client(server_certs: &[&[u8]]) -> Result<ClientConfig, Box<dyn Error>> {
-//     let mut certs = rustls::RootCertStore::empty();
-//     for cert in server_certs {
-//         certs.add(&rustls::Certificate(cert.to_vec()))?;
-//     }
-
-//     Ok(ClientConfig::with_root_certificates(certs))
-// }
+use quinn::{ClientConfig, ServerConfig, TransportConfig};
+use std::{error::Error, sync::Arc, time::Duration};
 
 /// Returns default server configuration along with its certificate.
 #[allow(clippy::field_reassign_with_default)] // https://github.com/rust-lang/rust-clippy/issues/6527
@@ -27,7 +13,10 @@ pub fn configure_server() -> Result<(ServerConfig, Vec<u8>), Box<dyn Error>> {
     let mut server_config = ServerConfig::with_single_cert(cert_chain, priv_key)?;
     Arc::get_mut(&mut server_config.transport)
         .unwrap()
-        .max_concurrent_uni_streams(0_u8.into());
+        .max_concurrent_uni_streams(0_u8.into())
+        .keep_alive_interval(Some(Duration::from_secs(1)))
+        // NOTE: removing this timeout might leave lingering connections around
+        .max_idle_timeout(None);
 
     Ok((server_config, cert_der))
 }
@@ -62,5 +51,12 @@ pub fn configure_client() -> ClientConfig {
         .with_custom_certificate_verifier(SkipServerVerification::new())
         .with_no_client_auth();
 
-    ClientConfig::new(Arc::new(crypto))
+    let mut transport = TransportConfig::default();
+    transport.keep_alive_interval(Some(Duration::from_secs(1)));
+    // NOTE: removing this timeout might leave lingering connections around
+    transport.max_idle_timeout(None);
+
+    let mut cli = ClientConfig::new(Arc::new(crypto));
+    cli.transport_config(Arc::new(transport));
+    cli
 }
